+++ osu.Game/Overlays/Settings/Sections/General/LanguageSettings.cs
@@ -20,7 +20,7 @@ namespace osu.Game.Overlays.Settings.Sections.General
                 {
                     LabelText = "以原语言显示歌曲信息",
                     Bindable = frameworkConfig.GetBindable<bool>(FrameworkSetting.ShowUnicode)
-                },
+                }
             };
         }
     }