+++ osu.Game/Screens/Multi/Components/RoomStatusInfo.cs
@@ -34,9 +34,9 @@ namespace osu.Game.Screens.Multi.Components
                 {
                     statusPart = new StatusPart
                     {
-                        Font = OsuFont.GetFont(weight: FontWeight.Bold, size: 18)
+                        Font = OsuFont.GetFont(weight: FontWeight.Bold, size: 17)
                     },
-                    endDatePart = new EndDatePart { Font = OsuFont.GetFont(size: 18) }
+                    endDatePart = new EndDatePart { Font = OsuFont.GetFont(size: 17) }
                 }
             };